Move and add geometry to walkable area builder issues
